// src/app/model/User.model.ts
export interface IUser {
    Id: string
    Name: string
    Surname: string
    Address: string
    ZIP: string
    City: string
    State: string
    Telephone: string
    CompanyName: string
    STN: string
    SSN: string
    Email: string
    Password: string
    Enabled: number
    Date: Date
    Role: string        
}